jhfv

+++ app/Http/Controllers/Dashboard/DcategoryController.php
@@ -81,10 +81,12 @@ class DcategoryController extends Controller
                 $imageurl->storeAs('public/categories', $file);
             $category->image = $file;
         }
-        $category->extras()->detach();
-        foreach ($request->extras as $type) {
-            $extras = Extra::where('type', $type)->get();
-            $category->extras()->attach($extras);
+        if($request->has('extras')){
+            $category->extras()->detach();
+            foreach ($request->extras as $type) {
+                $extras = Extra::where('type', $type)->get();
+                $category->extras()->attach($extras);
+            }
         }
